requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-escpos==3.1
pyserial==3.5
paho-mqtt==2.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.2.0
requests==2.31.0